+++ projects/demo-maps/src/app/route-components/route-example-maplibre/route-example-maplibre.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { RouteExampleMaplibreComponent } from './route-example-maplibre.component';
+
+describe('RouteExampleMaplibreComponent', () => {
+  let component: RouteExampleMaplibreComponent;
+  let fixture: ComponentFixture<RouteExampleMaplibreComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [ RouteExampleMaplibreComponent ]
+    })
+    .compileComponents();
+
+    fixture = TestBed.createComponent(RouteExampleMaplibreComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});