+++ CHANGELOG.md
@@ -19,6 +19,7 @@
   - New UKIS library export utilities for other libraries.
 
 * **@dlr-eoc/map-maplibre:**
+  - A example has been added to the demo-maps to show how to work with the new maplibre library.
   - New UKIS library for working with [maplibre](https://maplibre.org/) was added.
 
 * **@dlr-eoc/map-cesium:**